Fix some cast bugs

+++ train-system/src/main/java/hu/bme/mit/train/system/TrainSystem.java
@@ -34,7 +34,7 @@ public class TrainSystem {
   			public void run() {
 			    getController().followSpeed();
 			  }
-			}, 3*1000, 3*1000);
+			}, 3*(long)1000, 3*(long)1000);
 	}
 
 }
